\name{distM1}
\alias{distM1}
\title{An distance matrix constructed from L2 distance}
\description{This is the variable name for a distance matrix in the "Example" data. It is constructed from a sequence of 40 observations of dimension 10. The first 20 observations are considered historical observations. There is a change in mean at t = 10.
}

