% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path.R
\name{get_root_path}
\alias{get_root_path}
\title{Get the root data path}
\usage{
get_root_path(filename = "~/.gstream")
}
\arguments{
\item{filename}{chr, the file to read}
}
\value{
character root data path
}
\description{
Get the root data path
}
