% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path.R
\name{gstream_path}
\alias{gstream_path}
\title{Build a gstream data path}
\usage{
gstream_path(..., root = get_root_path())
}
\arguments{
\item{...}{character path segements passed to \code{file.path}}

\item{root}{chr, the root path}
}
\value{
a file path
}
\description{
Build a gstream data path
}
