% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usn.R
\name{list_usn}
\alias{list_usn}
\title{List USN files}
\usage{
list_usn(what = c("orig", "ordered", "raw")[2], ...)
}
\arguments{
\item{what}{chr one of "orig", "ordered" or "raw"}

\item{...}{other arguments for list.files}
}
\value{
chr vector of filenames
}
\description{
List USN files
}
