% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rapid.R
\name{mocha_get_metadata}
\alias{mocha_get_metadata}
\title{Get the metadata  for variables}
\usage{
mocha_get_metadata(x, vars)
}
\arguments{
\item{x}{ncdf4 object}

\item{vars}{chr, one or more variable names}
}
\value{
tibble of metadata for variable
}
\description{
Get the metadata  for variables
}
