% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rapid.R
\name{mocha_time_only_vars}
\alias{mocha_time_only_vars}
\title{Retrieve the variable names that depend only upon time}
\usage{
mocha_time_only_vars(x, dim = "time")
}
\arguments{
\item{x}{ncdf4 object}

\item{dim}{the name of the dim that contains time}
}
\value{
character vector of vars that deoend upon time only
}
\description{
Retrieve the variable names that depend only upon time
}
