% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rapid.R
\name{plot.rapid_mocha_time}
\alias{plot.rapid_mocha_time}
\title{Plot RAPID-MOCHA data time series}
\usage{
\method{plot}{rapid_mocha_time}(x = read_rapid_mocha(), smooth = TRUE, by = c("none", "day", "year"))
}
\arguments{
\item{x}{the mocha table}

\item{smooth}{logical, if TRUE add a smoothing line (for timeseries only)}

\item{by}{chr, one of 'none' (default timeseries), 'day' (ghost plot by day over the years)
or 'year' a box plot by year
#' @return a ggplot2 object}
}
\description{
Plot RAPID-MOCHA data time series
}
