% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{read_bathymetry}
\alias{read_bathymetry}
\title{Read the bathymetry map}
\usage{
read_bathymetry(
  filename = system.file("extdata/bathy.tif", package = "gstream")
)
}
\arguments{
\item{filename}{chr the name of the file}
}
\value{
stars object
}
\description{
Read the bathymetry map
}
