% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patch.R
\name{read_patch_bbs}
\alias{read_patch_bbs}
\title{Read the patches bounding boxes}
\usage{
read_patch_bbs(filename = gstream_path("patches/patches_bb.gpkg"))
}
\arguments{
\item{file}{chr, the name of the file to read}
}
\value{
sf POLYGON object with one or more patch polygons
}
\description{
Read the patches bounding boxes
}
