% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rapid.R
\name{read_rapid_mocha}
\alias{read_rapid_mocha}
\title{Read RAPID-MOCHA data}
\usage{
read_rapid_mocha(
  filename = gstream_path("rapid-mocha/mocha_mht_data_ERA5_v2020.nc"),
  by = "time"
)
}
\arguments{
\item{filename}{}

\item{by}{chr, one of 'time' (default) or 'depth'}
}
\value{
for \code{by = 'time'} then a tibble of data.  The table has an
added attribute called 'meta' that contains the column units and long name.
It will have the class "rapid_mocha_time"
}
\description{
Read RAPID-MOCHA data
}
